backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
python-multipart==0.0.6

# AI/ML Libraries
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
sentence-transformers==2.2.2
langchain==0.0.335
langchain-community==0.0.5

# Vector Database
pinecone-client==2.2.4

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Processing
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# OpenAI (optional)
openai==1.3.7

# Hugging Face
huggingface-hub==0.19.4

# HTTP Requests
requests==2.31.0
aiofiles==23.2.1

# CORS
fastapi[all]

# Development
black==23.11.0
pytest==7.4.3
jupyter==1.0.0